migrate for react-native 0.69
