+++ android/build.gradle
@@ -9,8 +9,8 @@ buildscript {
     dbFlowVersion = '4.2.4'
     buildToolsVersion = '31.0.0'
     kotlinVersion = '1.6.10'
-    gradlePluginVersion = '7.0.4'
-    gradleDownloadTaskVersion = '4.1.2'
+    gradlePluginVersion = '7.1.1'
+    gradleDownloadTaskVersion = '5.0.1'
     repositoryUrl = "file:${System.env.HOME}/.m2/repository/"
   }
   repositories {
