+++ android/expoview/build.gradle
@@ -181,7 +181,7 @@ android {
       "**/libglog.so",
       "**/libjscexecutor.so",
       "**/libfbjni.so",
-      "**/libfolly_json.so",
+      "**/libfolly_runtime.so",
       "**/libhermes.so",
       "**/libjsi.so",
     ]
@@ -462,7 +462,7 @@ task createNativeDepsDirectories {
 }
 
 task downloadBoost(dependsOn: createNativeDepsDirectories, type: Download) {
-  src("https://github.com/react-native-community/boost-for-react-native/releases/download/v${BOOST_VERSION.replace("_", ".")}-0/boost_${BOOST_VERSION}.tar.gz")
+  src("https://boostorg.jfrog.io/artifactory/main/release/${BOOST_VERSION.replace("_", ".")}/source/boost_${BOOST_VERSION}.tar.gz")
   onlyIfNewer(true)
   overwrite(false)
   dest(new File(downloadsDir, "boost_${BOOST_VERSION}.tar.gz"))
