+++ android/gradle.properties
@@ -13,3 +13,6 @@ android.useAndroidX=true
 android.enableJetifier=true
 android.useNewApkCreator=false
 expo.updates.createManifest=false
+
+# Override the react-native version where expo modules got from node_modules
+reactNativeVersion=0.69.0
