--- packages/expo/android/src/legacyReactNativeHostWrapper/expo/modules/ReactNativeHostWrapper.kt
@@ -1,9 +1,0 @@
-package expo.modules
-
-import android.app.Application
-import com.facebook.react.ReactNativeHost
-
-class ReactNativeHostWrapper(
-  application: Application,
-  host: ReactNativeHost
-) : ReactNativeHostWrapperBase(application, host)
